Module aws_sdk_iotevents::model

Modules

acknowledge_flow

See AcknowledgeFlow

action

See Action

alarm_action

See AlarmAction

alarm_capabilities

See AlarmCapabilities

alarm_event_actions

See AlarmEventActions

alarm_model_summary

See AlarmModelSummary

alarm_model_version_summary

See AlarmModelVersionSummary

alarm_notification

See AlarmNotification

alarm_rule

See AlarmRule

analysis_result

See AnalysisResult

analysis_result_location

See AnalysisResultLocation

asset_property_timestamp

See AssetPropertyTimestamp

asset_property_value

See AssetPropertyValue

asset_property_variant

See AssetPropertyVariant

attribute

See Attribute

clear_timer_action

See ClearTimerAction

detector_debug_option

See DetectorDebugOption

detector_model

See DetectorModel

detector_model_configuration

See DetectorModelConfiguration

detector_model_definition

See DetectorModelDefinition

detector_model_summary

See DetectorModelSummary

detector_model_version_summary

See DetectorModelVersionSummary

dynamo_d_bv2_action

See DynamoDBv2Action

dynamo_db_action

See DynamoDbAction

email_configuration

See EmailConfiguration

email_content

See EmailContent

email_recipients

See EmailRecipients

event

See Event

firehose_action

See FirehoseAction

initialization_configuration

See InitializationConfiguration

input

See Input

input_configuration

See InputConfiguration

input_definition

See InputDefinition

input_identifier

See InputIdentifier

input_summary

See InputSummary

iot_events_action

See IotEventsAction

iot_events_input_identifier

See IotEventsInputIdentifier

iot_site_wise_action

See IotSiteWiseAction

iot_site_wise_asset_model_property_identifier

See IotSiteWiseAssetModelPropertyIdentifier

iot_site_wise_input_identifier

See IotSiteWiseInputIdentifier

iot_topic_publish_action

See IotTopicPublishAction

lambda_action

See LambdaAction

logging_options

See LoggingOptions

notification_action

See NotificationAction

notification_target_actions

See NotificationTargetActions

on_enter_lifecycle

See OnEnterLifecycle

on_exit_lifecycle

See OnExitLifecycle

on_input_lifecycle

See OnInputLifecycle

payload

See Payload

recipient_detail

See RecipientDetail

reset_timer_action

See ResetTimerAction

routed_resource

See RoutedResource

set_timer_action

See SetTimerAction

set_variable_action

See SetVariableAction

simple_rule

See SimpleRule

sms_configuration

See SmsConfiguration

sns_topic_publish_action

See SnsTopicPublishAction

sqs_action

See SqsAction

sso_identity

See SsoIdentity

state

See State

tag

See Tag

transition_event

See TransitionEvent

Structs

AcknowledgeFlow

Specifies whether to get notified for alarm state changes.

Action

An action to be performed when the condition is TRUE.

AlarmAction

Specifies one of the following actions to receive notifications when the alarm state

AlarmCapabilities

Contains the configuration information of alarm state changes.

AlarmEventActions

Contains information about one or more alarm actions.

AlarmModelSummary

Contains a summary of an alarm model.

AlarmModelVersionSummary

Contains a summary of an alarm model version.

AlarmNotification

Contains information about one or more notification actions.

AlarmRule

Defines when your alarm is invoked.

AnalysisResult

Contains the result of the analysis.

AnalysisResultLocation

Contains information that you can use to locate the field in your detector model that the

AssetPropertyTimestamp

A structure that contains timestamp information. For more information, see TimeInNanos in the AWS IoT SiteWise API Reference.

AssetPropertyValue

A structure that contains value information. For more information, see AssetPropertyValue in the AWS IoT SiteWise API Reference.

AssetPropertyVariant

A structure that contains an asset property value. For more information, see Variant

Attribute

The attributes from the JSON payload that are made available by the input. Inputs are

ClearTimerAction

Information needed to clear the timer.

DetectorDebugOption

The detector model and the specific detectors (instances) for which the logging level is

DetectorModel

Information about the detector model.

DetectorModelConfiguration

Information about how the detector model is configured.

DetectorModelDefinition

Information that defines how a detector operates.

DetectorModelSummary

Information about the detector model.

DetectorModelVersionSummary

Information about the detector model version.

DynamoDBv2Action

Defines an action to write to the Amazon DynamoDB table that you created. The default action

DynamoDbAction

Defines an action to write to the Amazon DynamoDB table that you created. The standard action

EmailConfiguration

Contains the configuration information of email notifications.

EmailContent

Contains the subject and message of an email.

EmailRecipients

Contains the information of one or more recipients who receive the emails.

Event

Specifies the actions to be performed when the condition

FirehoseAction

Sends information about the detector model instance and the event that triggered the

InitializationConfiguration

Specifies the default alarm state.

Input

Information about the input.

InputConfiguration

Information about the configuration of an input.

InputDefinition

The definition of the input.

InputIdentifier

InputSummary

Information about the input.

IotEventsAction

Sends an AWS IoT Events input, passing in information about the detector model instance and the

IotEventsInputIdentifier

IotSiteWiseAction

Sends information about the detector model instance and the event that triggered the

IotSiteWiseAssetModelPropertyIdentifier

IotSiteWiseInputIdentifier

IotTopicPublishAction

Information required to publish the MQTT message through the AWS IoT message broker.

LambdaAction

Calls a Lambda function, passing in information about the detector model instance and the

LoggingOptions

The values of the AWS IoT Events logging options.

NotificationAction

Contains the notification settings of an alarm model.

NotificationTargetActions

Specifies an AWS Lambda function to manage alarm notifications.

OnEnterLifecycle

When entering this state, perform these actions if the condition

OnExitLifecycle

When exiting this state, perform these actions if the specified

OnInputLifecycle

Specifies the actions performed when the condition evaluates to TRUE.

Payload

Information needed to configure the payload.

RecipientDetail

The information that identifies the recipient.

ResetTimerAction

Information required to reset the timer. The timer is reset to the previously evaluated

RoutedResource

SetTimerAction

Information needed to set the timer.

SetVariableAction

Information about the variable and its new value.

SimpleRule

A rule that compares an input property value to a threshold value with a comparison operator.

SmsConfiguration

Contains the configuration information of SMS notifications.

SnsTopicPublishAction

Information required to publish the Amazon SNS message.

SqsAction

Sends information about the detector model instance and the event that triggered the

SsoIdentity

Contains information about your identity source in AWS Single Sign-On. For more information, see

State

Information that defines a state of a detector.

Tag

Metadata that can be used to manage the resource.

TransitionEvent

Specifies the actions performed and the next state entered when a condition

Enums

AlarmModelVersionStatus

AnalysisResultLevel

AnalysisStatus

ComparisonOperator

DetectorModelVersionStatus

EvaluationMethod

InputStatus

LoggingLevel

PayloadType