Module aws_sdk_iotevents::operation

Structs

CreateAlarmModel

Creates an alarm model to monitor an AWS IoT Events input attribute. You can use the alarm to get

CreateDetectorModel

Creates a detector model.

CreateInput

Creates an input.

DeleteAlarmModel

Deletes an alarm model. Any alarm instances that were created based on this alarm model

DeleteDetectorModel

Deletes a detector model. Any active instances of the detector model are also

DeleteInput

Deletes an input.

DescribeAlarmModel

Retrieves information about an alarm model. If you don't specify a value for the

DescribeDetectorModel

Describes a detector model. If the version parameter is not specified,

DescribeDetectorModelAnalysis

Retrieves runtime information about a detector model analysis.

DescribeInput

Describes an input.

DescribeLoggingOptions

Retrieves the current settings of the AWS IoT Events logging options.

GetDetectorModelAnalysisResults

Retrieves one or more analysis results of the detector model.

ListAlarmModelVersions

Lists all the versions of an alarm model. The operation returns only the metadata

ListAlarmModels

Lists the alarm models that you created. The operation returns only the metadata

ListDetectorModelVersions

Lists all the versions of a detector model. Only the metadata associated with each

ListDetectorModels

Lists the detector models you have created. Only the metadata associated with each

ListInputRoutings

ListInputs

Lists the inputs you have created.

ListTagsForResource

Lists the tags (metadata) you have assigned to the resource.

PutLoggingOptions

Sets or updates the AWS IoT Events logging options.

StartDetectorModelAnalysis

Performs an analysis of your detector model. For more information,

TagResource

Adds to or modifies the tags of the given resource. Tags are metadata that can be used to

UntagResource

Removes the given tags (metadata) from the resource.

UpdateAlarmModel

Updates an alarm model. Any alarms that were created based on the previous version are

UpdateDetectorModel

Updates a detector model. Detectors (instances) spawned by the previous version are

UpdateInput

Updates an input.