Struct aws_sdk_iotevents::output::DescribeAlarmModelOutput

#[non_exhaustive]
pub struct DescribeAlarmModelOutput {

    pub creation_time: Option<Instant>,
    pub alarm_model_arn: Option<String>,
    pub alarm_model_version: Option<String>,
    pub last_update_time: Option<Instant>,
    pub status: Option<AlarmModelVersionStatus>,
    pub status_message: Option<String>,
    pub alarm_model_name: Option<String>,
    pub alarm_model_description: Option<String>,
    pub role_arn: Option<String>,
    pub key: Option<String>,
    pub severity: Option<i32>,
    pub alarm_rule: Option<AlarmRule>,
    pub alarm_notification: Option<AlarmNotification>,
    pub alarm_event_actions: Option<AlarmEventActions>,
    pub alarm_capabilities: Option<AlarmCapabilities>,

}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

creation_time: Option<Instant>

The time the alarm model was created, in the Unix epoch format.

alarm_model_arn: Option<String>

The ARN of the alarm model. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

alarm_model_version: Option<String>

The version of the alarm model.

last_update_time: Option<Instant>

The time the alarm model was last updated, in the Unix epoch format.

status: Option<AlarmModelVersionStatus>

The status of the alarm model. The status can be one of the following values:

ACTIVE - The alarm model is active and it's ready to evaluate data.

ACTIVATING - AWS IoT Events is activating your alarm model. Activating an alarm model can take up to a few minutes.

INACTIVE - The alarm model is inactive, so it isn't ready to evaluate data. Check your alarm model information and update the alarm model.

FAILED - You couldn't create or update the alarm model. Check your alarm model information and try again.

status_message: Option<String>

Contains information about the status of the alarm model.

alarm_model_name: Option<String>

The name of the alarm model.

alarm_model_description: Option<String>

The description of the alarm model.

role_arn: Option<String>

The ARN of the IAM role that allows the alarm to perform actions and access AWS resources. For more information, see Amazon Resource Names (ARNs) in the AWS General Reference.

key: Option<String>

An input attribute used as a key to create an alarm. AWS IoT Events routes inputs associated with this key to the alarm.

severity: Option<i32>

A non-negative integer that reflects the severity level of the alarm.

alarm_rule: Option<AlarmRule>

Defines when your alarm is invoked.

alarm_notification: Option<AlarmNotification>

Contains information about one or more notification actions.

alarm_event_actions: Option<AlarmEventActions>

Contains information about one or more alarm actions.

alarm_capabilities: Option<AlarmCapabilities>

Contains the configuration information of alarm state changes.

Implementations

impl DescribeAlarmModelOutput

pub fn builder() -> Builder

Creates a new builder-style object to manufacture DescribeAlarmModelOutput